Document the conditional dependency check
Because the tiler compiles the operator for a value just once, if the
result of that operator is first used in a conditional block and then
used in an unconditional block (or a block that doesn't follow it),
that value will not be defined in all cases and the compiled code will
be incorrect.

I think I can prevent this by doing a check at template precompilation
time.